Divide and utilize JSX in separate page components, then refactor routes.js and store.js to utilize ABFactory via app.js
